Add project
Project was initially made in a private repo which documents my coding
journey, since restarting the Odin Project web development Foundations
curriculum in July 2023. Full commit history of this project can be seen
in the private repo. In future, I may make that repo public. Until then,
this repo serves the purpose of sharing this project in an easily
accessible way (without having to delve through a large file structure,
as would be the case in the coding journey repo).