Replace root-relative paths with relative paths
File paths were broken on GitHub Pages and so they had to be changed in
all applicable documnents. I discovered that root-relative file paths
don't work with GitHub Pages, because a new directory of the repo name is
created above the original directory/file structure of a project. In
other words, said directory takes the place of the directories and files
that were originally directly relative to the root directory. For
example, the position of the "recipe" directory within this project's
directory/file structure on my computer is as follows:

/recipes — it is immediately related to the project's root directory.

On GitHub Pages, its position becomes /odin-recipes/recipes — in this
setting, "odin-recipes" is the only directory immediately relative to the
root directory. All other directories and files that were originally
child to the root on the local filesystem are now child to "odin-recipes".
Therefore, in the context of the GitHub Pages service, /recipes
doesn't exist.